lambdas/requirements.txt
boto3==1.12.17
geohash2==1.1
marshmallow==3.5.1
psycopg2-binary==2.8.4
python-dotenv==0.12.0
pytz==2019.3
